requirements.txt
asgiref==3.5.0
coverage==6.3.2
CurrencyConverter==0.16.12
Django==4.0.4
django-cors-headers==3.13.0
django-modeltranslation==0.18.2
djangorestframework==3.13.1
freezegun==1.2.2
gunicorn==20.1.0
mysqlclient==2.1.1
protobuf==4.21.4
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
google-cloud-datastore=2.20.1